# Documents/LeFax Talents Dev Projects /backendsempos-master/api-doc/apidoc_v1.yaml
openapi: 3.0.2
info:
  title: Onebaz API
  version: '1.0'
  contact:
    name: Onebaz
    email: sempos@gmx.com
    url: www.onebaz.com
  description: API for the Onebaz back-end system
servers:
  - url: 'https://api.onebaz.com/v1/base'
  - url: 'https://api.onebaz.com/v1/{tenant}'
    variables:
      tenant:
        description: Tenant name
        default: lefax
        enum:
          - lefax
          - yellowshop
          - salaudbar
          - idealdallas
  - url: 'http://localhost:3000/v1/base'
  - url: 'http://localhost:3000/v1/{tenant}'
    variables:
      tenant:
        description: Tenant name
        default: velocity
        enum:
          - velocity
          - yellowshop
          - salaudbar
          - idealdallas
paths:
...
  /customers:
    get:
      summary: Get all customers
      tags:
        - Customers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  count:
                    type: integer
                  currentPage:
                    type: integer
                  pages:
                    type: integer
                  prev:
                    type: string
                    format: uri
                    nullable: true
                  next:
                    type: string
                    format: uri
                    nullable: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Customer'
      operationId: get-customers
      security:
        - Tenant User Api Key: []
      parameters:
        - schema:
            type: integer
          in: query
          name: page
        - schema:
            type: integer
          in: query
          name: limit
    post:
      summary: Create a new customer
      operationId: post-customers
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Customer'
                  url:
                    type: string
                    format: uri
      security:
        - Tenant User Api Key: []
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                type:
                  type: string
                  enum:
                    - individual
                    - corporate
                email:
                  type: string
                phones:
                  type: array
                  items:
                    type: string
                fax:
                  type: string
                customers_groups:
                  type: array
                  items:
                    type: string
                sendMarketingEmails:
                  type: boolean
                street:
                  type: string
                city:
                  type: string
                post_code:
                  type: string
                state:
                  type: string
                country:
                  type: string
                birth_date:
                  type: string
                gender:
                  type: string
                  enum:
                    - female
                    - male
                website:
                  type: string
                  format: uri
                enableLoyalty:
                  type: boolean
              required:
                - name
                - type
      tags:
        - Customers
    parameters: []
  /customers/verify:
    get:
      summary: Check name availability for customer
      tags:
        - Customers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
      operationId: get-customers-verify
      security:
        - Tenant User Api Key: []
      parameters:
        - schema:
            type: string
          in: query
          name: name
          required: true
  '/customers/{id}':
    parameters:
      - schema:
          type: string
        name: id
        in: path
        required: true
    get:
      summary: Get a customer by id
      tags:
        - Customers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
      operationId: get-customers-id
      security:
        - Tenant User Api Key: []
    patch:
      summary: Update a customer
      operationId: patch-customers-id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
      security:
        - Tenant User Api Key: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                type:
                  type: string
                  enum:
                    - person
                    - company
                email:
                  type: string
                phones_to_add:
                  type: array
                  items:
                    type: string
                phones_to_remove:
                  type: array
                  items:
                    type: string
                fax:
                  type: string
                customers_groups_to_add:
                  type: array
                  items:
                    type: string
                customers_groups_to_remove:
                  type: array
                  items:
                    type: string
                sendMarketingEmails:
                  type: boolean
                street:
                  type: string
                city:
                  type: string
                post_code:
                  type: string
                state:
                  type: string
                country:
                  type: string
                birth_date:
                  type: string
                gender:
                  type: string
                  enum:
                    - female
                    - male
                website:
                  type: string
                  format: uri
                enableLoyalty:
                  type: boolean
      tags:
        - Customers
    delete:
      summary: Delete a customer
      operationId: delete-customers-id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
      security:
        - Tenant User Api Key: []
      tags:
        - Customers
...
  '/customersgroups/{id}/customers':
    parameters:
      - schema:
          type: string
        name: id
        in: path
        required: true
    get:
      summary: Get customers of a group
      tags:
        - CustomersGroups
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  count:
                    type: integer
                  currentPage:
                    type: integer
                  pages:
                    type: integer
                  prev:
                    type: string
                    nullable: true
                  next:
                    type: string
                    nullable: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Customer'
      operationId: get-customersgroups-id-customers
      security:
        - Tenant User Api Key: []
      parameters:
        - schema:
            type: integer
          in: query
          name: page
        - schema:
            type: integer
          in: query
          name: limit
  /customers/search:
    get:
      summary: Get a customer by code
      tags:
        - Customers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
      operationId: get-customers-search
      parameters:
        - schema:
            type: string
          in: query
          name: code
          required: true
      security:
        - Tenant User Api Key: []
...
  /purchases:
    get:
      summary: Get purchases
      tags:
        - Purchases
      responses:
        '200':
          $ref: '#/components/responses/GetPurchases'
      operationId: get-purchases
      security:
        - Tenant User Api Key: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
    post:
      summary: Add a purchase
      operationId: post-purchases
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/GetPurchase'
                  url:
                    type: string
                    format: uri
      security:
        - Tenant User Api Key: []
      requestBody:
        $ref: '#/components/requestBodies/AddPurchase'
      tags:
        - Purchases
  '/purchases/{id}':
    parameters:
      - schema:
          type: string
        name: id
        in: path
        required: true
    get:
      summary: Get a purchase by id
      tags:
        - Purchases
      responses:
        '200':
          $ref: '#/components/responses/GetPurchase'
      operationId: get-purchases-id
      security:
        - Tenant User Api Key: []
    patch:
      summary: Update a purchase
      operationId: patch-purchases-id
      responses:
        '200':
          $ref: '#/components/responses/GetPurchase'
      security:
        - Tenant User Api Key: []
      requestBody:
        $ref: '#/components/requestBodies/UpdatePurchase'
      tags:
        - Purchases
...
components:
  schemas:
...
    Customer:
      title: Customer
      type: object
      properties:
        id:
          type: string
        infos:
          type: object
          properties:
            name:
              type: string
            type:
              type: string
              enum:
                - person
                - company
        contact:
          type: object
          properties:
            email:
              type: string
              format: email
            phones:
              type: array
              items:
                type: string
            fax:
              type: string
            customers_groups:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
                description:
                  type: string
            sendMarketingEmails:
              type: boolean
        details:
          type: object
          properties:
            postal_address:
              type: object
              properties:
                street:
                  type: string
                city:
                  type: string
                post_code:
                  type: string
                state:
                  type: string
                country:
                  type: string
            additional_information:
              type: object
              properties:
                customer_code:
                  type: string
                birth_date:
                  type: string
                gender:
                  type: string
                  enum:
                    - female
                    - male
                website:
                  type: string
                  format: uri
                enableLoyalty:
                  type: boolean
...
    GetPurchase:
      title: GetPurchase
      type: object
      properties:
        _id:
          type: string
        tenant:
          type: string
        employee:
          type: string
        infos:
          type: object
          properties:
            purchase_number:
              type: string
            expected_arrival_date:
              type: string
              format: date
            vendor:
              type: string
            branch_to_be_shipped:
              type: string
            invoice_number:
              type: string
            billing_date:
              type: string
              format: date
            status:
              type: string
              enum:
                - '''waiting_for_approval'''
                - '''waiting_for_transfer'''
                - '''partially_completed'''
                - '''completed'''
            description:
              type: string
        products:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              status:
                type: object
                properties:
                  waiting:
                    type: boolean
                    default: true
                  received:
                    type: number
                    default: 0
                  not_received:
                    type: number
                    default: 0
                  damaged:
                    type: number
                    default: 0
              quantity:
                type: number
              unit_price:
                type: number
              tax:
                type: number
              sub_total:
                type: string
        cart_discount:
          type: number
        total_price:
          type: number
    AddPurchase:
      title: AddPurchase
      type: object
      properties:
        '':
          type: object
          properties:
            purchase_number:
              type: string
            expected_arrival_date:
              type: string
              format: date
            vendor:
              type: string
            branch_to_be_shipped:
              type: string
            invoice_number:
              type: string
            billing_date:
              type: string
              format: date
            status:
              type: string
              enum:
                - '''waiting_for_approval'''
                - '''waiting_for_transfer'''
            description:
              type: string
          required:
            - branch_to_be_shipped
        products:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              quantity:
                type: number
              unit_price:
                type: number
              tax:
                type: number
        cart_discount:
          type: number
    UpdatePurchase:
      title: UpdatePurchase
      type: object
      properties:
        '':
          type: object
          properties:
            purchase_number:
              type: string
            expected_arrival_date:
              type: string
              format: date
            vendor:
              type: string
            branch_to_be_shipped:
              type: string
            invoice_number:
              type: string
            billing_date:
              type: string
              format: date
            status:
              type: string
              enum:
                - waiting_for_approval
                - waiting_for_transfer
            description:
              type: string
          required:
            - branch_to_be_shipped
        products:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              status:
                type: object
                properties:
                  waiting:
                    type: boolean
                    default: true
                  received:
                    type: number
                    default: 0
                  not_received:
                    type: number
                    default: 0
                  damaged:
                    type: number
                    default: 0
              quantity:
                type: number
              unit_price:
                type: number
              tax:
                type: number
        cart_discount:
          type: number
...
  securitySchemes:
    Admin Api Key:
      name: admin_api_key
      type: apiKey
      in: header
    Tenant User Api Key:
      name: tenant_user_api_key
      type: apiKey
      in: header
  requestBodies:
...
    AddPurchase:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AddPurchase'
    UpdatePurchase:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UpdatePurchase'
...
  parameters:
...
    page:
      name: page
      in: query
      required: false
      schema:
        type: integer
    limit:
      name: limit
      in: query
      required: false
      schema:
        type: string
...
  responses:
...
    GetPurchases:
      description: Example response
      content:
        application/json:
          schema:
            type: object
            properties:
              total:
                type: integer
              count:
                type: integer
              currentPage:
                type: integer
              pages:
                type: integer
              prevPage:
                type: string
                format: uri
                nullable: true
              nextPage:
                type: string
                format: uri
                nullable: true
              data:
                $ref: '#/components/schemas/GetPurchase'
    GetPurchase:
      description: Example response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GetPurchase'
...